be-25.4.29661.29727

+++ BaseApp/Source/Base Application/Local/BankMgt/Statement/CODAWriteStatements.Codeunit.al
@@ -70,15 +70,18 @@ codeunit 2000041 "CODA Write Statements"
                     CodedBankStnt."Statement No." := CodBankStmtSrcLine."Statement No.";
                     CodedBankStnt."Statement Date" := CodBankStmtSrcLine."Transaction Date";
                     if not CodedBankStnt.Insert(true) then
-                        if Confirm(
-                            StrSubstNo(Text000,
-                            CodedBankStnt.TableCaption(), CodBankStmtSrcLine."Bank Account No.", CodBankStmtSrcLine."Statement No."))
-                        then begin
-                            CodBankStmtLine.Reset();
-                            CodBankStmtLine.SetRange("Bank Account No.", CodBankStmtSrcLine."Bank Account No.");
-                            CodBankStmtLine.SetRange("Statement No.", CodBankStmtSrcLine."Statement No.");
-                            CodBankStmtLine.DeleteAll(true);
-                            CodedBankStnt.Modify(true);
+                        if GuiAllowed then begin
+                            if Confirm(
+                                StrSubstNo(Text000,
+                                CodedBankStnt.TableCaption(), CodBankStmtSrcLine."Bank Account No.", CodBankStmtSrcLine."Statement No."))
+                            then begin
+                                CodBankStmtLine.Reset();
+                                CodBankStmtLine.SetRange("Bank Account No.", CodBankStmtSrcLine."Bank Account No.");
+                                CodBankStmtLine.SetRange("Statement No.", CodBankStmtSrcLine."Statement No.");
+                                CodBankStmtLine.DeleteAll(true);
+                                CodedBankStnt.Modify(true);
+                            end else
+                                exit(false);
                         end else
                             exit(false);
                     Clear(LastCodBankStmtSrcLine);
